API

Types

General basis types. Specialized bases can be found in the section API: Quantum-systems.

QuantumOpticsBase.Basis — Type

Abstract base class for all specialized bases.

The Basis class is meant to specify a basis of the Hilbert space of the studied system. Besides basis specific information all subclasses must implement a shape variable which indicates the dimension of the used Hilbert space. For a spin-1/2 Hilbert space this would be the vector Int[2]. A system composed of two spins would then have a shape vector Int[2 2].

Composite systems can be defined with help of the CompositeBasis class.

QuantumOpticsBase.GenericBasis — Type

GenericBasis(N)

A general purpose basis of dimension N.

Should only be used rarely since it defeats the purpose of checking that the bases of state vectors and operators are correct for algebraic operations. The preferred way is to specify special bases for different systems.

QuantumOpticsBase.CompositeBasis — Type

CompositeBasis(b1, b2...)

Basis for composite Hilbert spaces.

Stores the subbases in a vector and creates the shape vector directly from the shape vectors of these subbases. Instead of creating a CompositeBasis directly tensor(b1, b2...) or b1 ⊗ b2 ⊗ … can be used.

States

QuantumOpticsBase.StateVector — Type

Abstract base class for Bra and Ket states.

The state vector class stores the coefficients of an abstract state in respect to a certain basis. These coefficients are stored in the data field and the basis is defined in the basis field.

QuantumOpticsBase.Bra — Type

Bra(b::Basis[, data])

Bra state defined by coefficients in respect to the basis.

QuantumOpticsBase.Ket — Type

Ket(b::Basis[, data])

Ket state defined by coefficients in respect to the given basis.

General purpose QuantumOpticsBase. A few more specialized operators are implemented in API: Quantum-systems.

QuantumOpticsBase.AbstractOperator — Type

Abstract base class for all operators.

All deriving operator classes have to define the fields basis_l and basis_r defining the left and right side bases.

For fast time evolution also at least the function mul!(result::Ket,op::AbstractOperator,x::Ket,alpha,beta) should be implemented. Many other generic multiplication functions can be defined in terms of this function and are provided automatically.

QuantumOpticsBase.DataOperator — Type

Abstract type for operators with a data field.

This is an abstract type for operators that have a direct matrix representation stored in their .data field.

QuantumOpticsBase.Operator — Type

Operator{BL,BR,T} <: DataOperator{BL,BR}

Operator type that stores the representation of an operator on the Hilbert spaces given by BL and BR (e.g. a Matrix).

QuantumOpticsBase.DenseOperator — Function

DenseOperator(b1[, b2, data])

Dense array implementation of Operator. Converts any given data to a dense Matrix.

QuantumOpticsBase.SparseOperator — Function

SparseOperator(b1[, b2, data])

Sparse array implementation of Operator.

The matrix is stored as the julia built-in type SparseMatrixCSC in the data field.

QuantumOpticsBase.LazyTensor — Type

LazyTensor(b1[, b2], indices, operators[, factor=1])

Lazy implementation of a tensor product of operators.

The suboperators are stored in the operators field. The indices field specifies in which subsystem the corresponding operator lives. Additionally, a complex factor is stored in the factor field which allows for fast multiplication with numbers.

QuantumOpticsBase.LazySum — Type

LazySum([factors,] operators)

Lazy evaluation of sums of operators.

All operators have to be given in respect to the same bases. The field factors accounts for an additional multiplicative factor for each operator stored in the operators field.

QuantumOpticsBase.LazyProduct — Type

LazyProduct(operators[, factor=1])
LazyProduct(op1, op2...)

Lazy evaluation of products of operators.

The factors of the product are stored in the operators field. Additionally a complex factor is stored in the factor field which allows for fast multiplication with numbers.

Super operators:

QuantumOpticsBase.SuperOperator — Type

SuperOperator <: AbstractSuperOperator

SuperOperator stored as representation, e.g. as a Matrix.

QuantumOpticsBase.DenseSuperOperator — Function

DenseSuperOperator(b1[, b2, data])

SuperOperator stored as dense matrix.

QuantumOpticsBase.SparseSuperOperator — Function

SparseSuperOperator(b1[, b2, data])

SuperOperator stored as sparse matrix.

Functions

Functions to generate general states, operators and super-operators

QuantumOpticsBase.basisstate — Function

basisstate(b, index; sparse=false, dType=ComplexF64)

Basis vector specified by index as ket state.

For a composite system index can be a vector which then creates a tensor product state $|i_1⟩⊗|i_2⟩⊗…⊗|i_n⟩$ of the corresponding basis states.

basisstate(b::ManyBodyBasis, occupation::Vector{Int})

Return a ket state where the system is in the state specified by the given occupation numbers.

QuantumOpticsBase.identityoperator — Function

identityoperator(a::Basis[, b::Basis])

Return an identityoperator in the given bases.

QuantumOpticsBase.diagonaloperator — Function

diagonaloperator(b::Basis)

Create a diagonal operator of type SparseOperator.

QuantumOpticsBase.randoperator — Function

randoperator(b1[, b2])

Calculate a random unnormalized dense operator.

QuantumOpticsBase.spre — Function

spre(op)

Create a super-operator equivalent for right side operator multiplication.

For operators $A$, $B$ the relation

\[ \mathrm{spre}(A) B = A B\]

holds. op can be a dense or a sparse operator.

QuantumOpticsBase.spost — Function

Create a super-operator equivalent for left side operator multiplication.

For operators $A$, $B$ the relation

\[ \mathrm{spost}(A) B = B A\]

holds. op can be a dense or a sparse operator.

QuantumOpticsBase.liouvillian — Function

liouvillian(H, J; rates, Jdagger)

Create a super-operator equivalent to the master equation so that $\dot ρ = S ρ$.

The super-operator $S$ is defined by

\[S ρ = -\frac{i}{ħ} [H, ρ] + \sum_i J_i ρ J_i^† - \frac{1}{2} J_i^† J_i ρ - \frac{1}{2} ρ J_i^† J_i\]

Arguments

H: Hamiltonian.

J: Vector containing the jump operators.

rates: Vector or matrix specifying the coefficients for the jump operators.

Jdagger: Vector containing the hermitian conjugates of the jump operators. If they are not given they are calculated automatically.

As far as it makes sense the same functions are implemented for bases, states, operators and superQuantumOpticsBase.

QuantumOpticsBase.samebases — Function

samebases(a, b)

Test if two objects have the same bases.

QuantumOpticsBase.check_samebases — Function

check_samebases(a, b)

Throw an IncompatibleBases error if the objects don't have the same bases.

QuantumOpticsBase.multiplicable — Function

multiplicable(a, b)

Check if two objects are multiplicable.

QuantumOpticsBase.check_multiplicable — Function

check_multiplicable(a, b)

Throw an IncompatibleBases error if the objects are not multiplicable.

QuantumOpticsBase.basis — Function

basis(a)

Return the basis of an object.

If it's ambiguous, e.g. if an operator has a different left and right basis, an IncompatibleBases error is thrown.

QuantumOpticsBase.dagger — Function

dagger(x)

Hermitian conjugate.

QuantumOpticsBase.tensor — Function

tensor(x, y, z...)

Tensor product of the given objects. Alternatively, the unicode symbol ⊗ (\otimes) can be used.

tensor(x::Basis, y::Basis, z::Basis...)

Create a CompositeBasis from the given bases.

Any given CompositeBasis is expanded so that the resulting CompositeBasis never contains another CompositeBasis.

tensor(x::Ket, y::Ket, z::Ket...)

Tensor product $|x⟩⊗|y⟩⊗|z⟩⊗…$ of the given states.

tensor(x::AbstractOperator, y::AbstractOperator, z::AbstractOperator...)

Tensor product $\hat{x}⊗\hat{y}⊗\hat{z}⊗…$ of the given operators.

tensor(x::Ket, y::Bra)

Outer product $|x⟩⟨y|$ of the given states.

QuantumOpticsBase.projector — Method

projector(a::Ket, b::Bra)

Projection operator $|a⟩⟨b|$.

QuantumOpticsBase.projector — Method

projector(a::Ket)

Projection operator $|a⟩⟨a|$.

QuantumOpticsBase.projector — Method

projector(a::Bra)

Projection operator $|a⟩⟨a|$.

QuantumOpticsBase.dm — Function

dm(a::StateVector)

Create density matrix $|a⟩⟨a|$. Same as projector(a).

LinearAlgebra.norm — Method

norm(x::StateVector)

Norm of the given bra or ket state.

LinearAlgebra.tr — Function

tr(x::AbstractOperator)

Trace of the given operator.

QuantumOpticsBase.ptrace — Function

ptrace(a, indices)

Partial trace of the given basis, state or operator.

The indices argument, which can be a single integer or a vector of integers, specifies which subsystems are traced out. The number of indices has to be smaller than the number of subsystems, i.e. it is not allowed to perform a full trace.

LinearAlgebra.normalize — Method

normalize(x::StateVector)

Return the normalized state so that norm(x) is one.

LinearAlgebra.normalize — Method

normalize(op)

Return the normalized operator so that its tr(op) is one.

LinearAlgebra.normalize! — Method

normalize!(x::StateVector)

In-place normalization of the given bra or ket so that norm(x) is one.

LinearAlgebra.normalize! — Method

normalize!(op)

In-place normalization of the given operator so that its tr(x) is one.

QuantumOpticsBase.expect — Function

expect(op, state)

Expectation value of the given operator op for the specified state.

state can either be a (density) operator or a ket.

expect(index, op, state)

If an index is given, it assumes that op is defined in the subsystem specified by this number.

QuantumOpticsBase.variance — Function

variance(op, state)

Variance of the given operator op for the specified state.

state can either be a (density) operator or a ket.

variance(index, op, state)

If an index is given, it assumes that op is defined in the subsystem specified by this number

QuantumOpticsBase.embed — Function

embed(basis1[, basis2], indices::Vector, operators::Vector)

Tensor product of operators where missing indices are filled up with identity operators.

embed(basis1[, basis2], indices::Vector, op::AbstractOperator)

Embed operator acting on a joint Hilbert space where missing indices are filled up with identity operators.

embed(basis1[, basis2], operators::Dict)

operators is a dictionary Dict{Vector{Int}, AbstractOperator}. The integer vector specifies in which subsystems the corresponding operator is defined.

QuantumOpticsBase.permutesystems — Function

permutesystems(a, perm)

Change the ordering of the subsystems of the given object.

For a permutation vector [2,1,3] and a given object with basis [b1, b2, b3] this function results in [b2, b1, b3].

Base.exp — Method

exp(op::AbstractOperator)

Operator exponential.

Missing docstring.

Missing docstring for QuantumOpticsBase.mul!(Y::DataOperator,A::AbstractOperator,B::DataOperator,alpha,beta). Check Documenter's build log for details.

Missing docstring.

Missing docstring for QuantumOpticsBase.mul!(Y::StateVector,A::AbstractOperator,B::StateVector,alpha,beta). Check Documenter's build log for details.

Conversion of operators

QuantumOpticsBase.dense — Function

dense(op::AbstractOperator)

Convert an arbitrary Operator into a DenseOperator.

SparseArrays.sparse — Method

sparse(op::AbstractOperator)

Convert an arbitrary operator into a SparseOperator.

Exceptions

QuantumOpticsBase.IncompatibleBases — Type

Exception that should be raised for an illegal algebraic operation.

Quantum systems

Fock

QuantumOpticsBase.FockBasis — Type

FockBasis(N,offset=0)

Basis for a Fock space where N specifies a cutoff, i.e. what the highest included fock state is. Similarly, the offset defines the lowest included fock state (default is 0). Note that the dimension of this basis is N+1-offset.

QuantumOpticsBase.number — Method

number(b::FockBasis)

Number operator for the specified Fock space.

QuantumOpticsBase.destroy — Method

destroy(b::FockBasis)

Annihilation operator for the specified Fock space.

QuantumOpticsBase.create — Method

create(b::FockBasis)

Creation operator for the specified Fock space.

QuantumOpticsBase.displace — Function

displace(b::FockBasis, alpha)

Displacement operator $D(α)$ for the specified Fock space.

QuantumOpticsBase.fockstate — Function

fockstate(b::FockBasis, n)

Fock state $|n⟩$ for the specified Fock space.

QuantumOpticsBase.coherentstate — Function

coherentstate(b::FockBasis, alpha)

Coherent state $|α⟩$ for the specified Fock space.

QuantumOpticsBase.coherentstate! — Function

coherentstate!(ket::Ket, b::FockBasis, alpha)

Inplace creation of coherent state $|α⟩$ for the specified Fock space.

N-level

QuantumOpticsBase.NLevelBasis — Type

NLevelBasis(N)

Basis for a system consisting of N states.

QuantumOpticsBase.transition — Method

transition(b::NLevelBasis, to::Int, from::Int)

Transition operator $|\mathrm{to}⟩⟨\mathrm{from}|$.

QuantumOpticsBase.nlevelstate — Function

nlevelstate(b::NLevelBasis, n::Int)

State where the system is completely in the n-th level.

Spin

QuantumOpticsBase.SpinBasis — Type

SpinBasis(n)

Basis for spin-n particles.

The basis can be created for arbitrary spinnumbers by using a rational number, e.g. SpinBasis(3//2). The Pauli operators are defined for all possible spin numbers.

QuantumOpticsBase.sigmax — Function

sigmax(b::SpinBasis)

Pauli $σ_x$ operator for the given Spin basis.

QuantumOpticsBase.sigmay — Function

sigmay(b::SpinBasis)

Pauli $σ_y$ operator for the given Spin basis.

QuantumOpticsBase.sigmaz — Function

sigmaz(b::SpinBasis)

Pauli $σ_z$ operator for the given Spin basis.

QuantumOpticsBase.sigmap — Function

sigmap(b::SpinBasis)

Raising operator $σ_+$ for the given Spin basis.

QuantumOpticsBase.sigmam — Function

sigmam(b::SpinBasis)

Lowering operator $σ_-$ for the given Spin basis.

QuantumOpticsBase.spinup — Function

spinup(b::SpinBasis)

Spin up state for the given Spin basis.

QuantumOpticsBase.spindown — Function

spindown(b::SpinBasis)

Spin down state for the given Spin basis.

Particle

QuantumOpticsBase.PositionBasis — Type

PositionBasis(xmin, xmax, Npoints)
PositionBasis(b::MomentumBasis)

Basis for a particle in real space.

For simplicity periodic boundaries are assumed which means that the rightmost point defined by xmax is not included in the basis but is defined to be the same as xmin.

When a MomentumBasis is given as argument the exact values of $x_{min}$ and $x_{max}$ are due to the periodic boundary conditions more or less arbitrary and are chosen to be $-\pi/dp$ and $\pi/dp$ with $dp=(p_{max}-p_{min})/N$.

QuantumOpticsBase.MomentumBasis — Type

MomentumBasis(pmin, pmax, Npoints)
MomentumBasis(b::PositionBasis)

Basis for a particle in momentum space.

For simplicity periodic boundaries are assumed which means that pmax is not included in the basis but is defined to be the same as pmin.

When a PositionBasis is given as argument the exact values of $p_{min}$ and $p_{max}$ are due to the periodic boundary conditions more or less arbitrary and are chosen to be $-\pi/dx$ and $\pi/dx$ with $dx=(x_{max}-x_{min})/N$.

QuantumOpticsBase.spacing — Function

spacing(b::PositionBasis)

Difference between two adjacent points of the real space basis.

spacing(b::MomentumBasis)

Momentum difference between two adjacent points of the momentum basis.

QuantumOpticsBase.samplepoints — Function

samplepoints(b::PositionBasis)

x values of the real space basis.

samplepoints(b::MomentumBasis)

p values of the momentum basis.

Base.position — Method

position(b::PositionBasis)

Position operator in real space.

Base.position — Method

position(b:MomentumBasis)

Position operator in momentum space.

QuantumOpticsBase.momentum — Method

momentum(b::PositionBasis)

Momentum operator in real space.

QuantumOpticsBase.momentum — Method

momentum(b:MomentumBasis)

Momentum operator in momentum space.

QuantumOpticsBase.potentialoperator — Function

potentialoperator(b::PositionBasis, V(x))

Operator representing a potential $V(x)$ in real space.

potentialoperator(b::MomentumBasis, V(x))

Operator representing a potential $V(x)$ in momentum space.

potentialoperator(b::CompositeBasis, V(x, y, z, ...))

Operator representing a potential $V$ in more than one dimension.

Arguments

b: Composite basis consisting purely either of PositionBasis or MomentumBasis. Note, that calling this with a composite basis in momentum space might consume a large amount of memory.

V: Function describing the potential. ATTENTION: The number of arguments accepted by V must match the spatial dimension. Furthermore, the order of the arguments has to match that of the order of the tensor product of bases (e.g. if b=bx⊗by⊗bz, then V(x,y,z)).

QuantumOpticsBase.gaussianstate — Function

gaussianstate(b::PositionBasis, x0, p0, sigma)
gaussianstate(b::MomentumBasis, x0, p0, sigma)

Create a Gaussian state around x0 andp0 with width sigma.

In real space the gaussian state is defined as

\[\Psi(x) = \frac{1}{\pi^{1/4}\sqrt{\sigma}} e^{i p_0 (x-\frac{x_0}{2}) - \frac{(x-x_0)^2}{2 \sigma^2}}\]

and is connected to the momentum space definition

\[\Psi(p) = \frac{\sqrt{\sigma}}{\pi^{1/4}} e^{-i x_0 (p-\frac{p_0}{2}) - \frac{1}{2}(p-p_0)^2 \sigma^2}\]

via a Fourier-transformation

\[\Psi(p) = \frac{1}{\sqrt{2\pi}} \int_{-\infty}^{\infty} e^{-ipx}\Psi(x) \mathrm{d}x\]

The state has the properties

$⟨p⟩ = p_0$

$⟨x⟩ = x_0$

$\mathrm{Var}(x) = \frac{σ^2}{2}$

$\mathrm{Var}(p) = \frac{1}{2 σ^2}$

Due to the numerically necessary discretization additional scaling factors $\sqrt{Δx}$ and $\sqrt{Δp}$ are used so that $\langle x_i|Ψ\rangle = \sqrt{Δ x} Ψ(x_i)$ and $\langle p_i|Ψ\rangle = \sqrt{Δ p} Ψ(p_i)$ so that the resulting Ket state is normalized.

QuantumOpticsBase.FFTOperator — Type

FFTOperator

Abstract type for all implementations of FFT operators.

QuantumOpticsBase.FFTOperators — Type

FFTOperators

Operator performing a fast fourier transformation when multiplied with a state that is a Ket or an Operator.

QuantumOpticsBase.FFTKets — Type

FFTKets

Operator that can only perform fast fourier transformations on Kets. This is much more memory efficient when only working with Kets.

QuantumOpticsBase.transform — Function

transform(b1::MomentumBasis, b2::PositionBasis)
transform(b1::PositionBasis, b2::MomentumBasis)

Transformation operator between position basis and momentum basis.

transform(b1::CompositeBasis, b2::CompositeBasis)

Transformation operator between two composite bases. Each of the bases has to contain bases of type PositionBasis and the other one a corresponding MomentumBasis.

transform(b1::PositionBasis, b2::FockBasis; x0=1)
transform(b1::FockBasis, b2::PositionBasis; x0=1)

Transformation operator between position basis and fock basis.

The coefficients are connected via the relation

\[ψ(x_i) = \sum_{n=0}^N ⟨x_i|n⟩ ψ_n\]

where $⟨x_i|n⟩$ is the value of the n-th eigenstate of a particle in a harmonic trap potential at position $x$, i.e.:

\[⟨x_i|n⟩ = π^{-\frac{1}{4}} \frac{e^{-\frac{1}{2}\left(\frac{x}{x_0}\right)^2}}{\sqrt{x_0}} \frac{1}{\sqrt{2^n n!}} H_n\left(\frac{x}{x_0}\right)\]

Subspace bases

QuantumOpticsBase.SubspaceBasis — Type

SubspaceBasis(basisstates)

A basis describing a subspace embedded a higher dimensional Hilbert space.

QuantumOpticsBase.orthonormalize — Function

orthonormalize(b::SubspaceBasis)

Orthonormalize the basis states of the given SubspaceBasis

A modified Gram-Schmidt process is used.

QuantumOpticsBase.projector — Method

projector(b1, b2; sparse=false)

Projection operator between subspaces and superspaces or between two subspaces.

Many-body

QuantumOpticsBase.ManyBodyBasis — Type

ManyBodyBasis(b, occupations)

Basis for a many body system.

The basis has to know the associated one-body basis b and which occupation states should be included. The occupations_hash is used to speed up checking if two many-body bases are equal.

QuantumOpticsBase.fermionstates — Function

fermionstates(Nmodes, Nparticles)
fermionstates(b, Nparticles)

Generate all fermionic occupation states for N-particles in M-modes. Nparticles can be a vector to define a Hilbert space with variable particle number.

QuantumOpticsBase.bosonstates — Function

bosonstates(Nmodes, Nparticles)
bosonstates(b, Nparticles)

Generate all bosonic occupation states for N-particles in M-modes. Nparticles can be a vector to define a Hilbert space with variable particle number.

QuantumOpticsBase.number — Method

number(b::ManyBodyBasis, index)

Particle number operator for the i-th mode of the many-body basis b.

QuantumOpticsBase.number — Method

number(b::ManyBodyBasis)

Total particle number operator.

QuantumOpticsBase.destroy — Method

destroy(b::ManyBodyBasis, index)

Annihilation operator for the i-th mode of the many-body basis b.

QuantumOpticsBase.create — Method

create(b::ManyBodyBasis, index)

Creation operator for the i-th mode of the many-body basis b.

QuantumOpticsBase.transition — Method

transition(b::ManyBodyBasis, to::Int, from::Int)

Operator $|\mathrm{to}⟩⟨\mathrm{from}|$ transferring particles between modes.

QuantumOpticsBase.manybodyoperator — Function

manybodyoperator(b::ManyBodyBasis, op)

Create the many-body operator from the given one-body operator op.

The given operator can either be a one-body operator or a two-body interaction. Higher order interactions are at the moment not implemented.

The mathematical formalism for the one-body case is described by

\[X = \sum_{ij} a_i^† a_j ⟨u_i| x | u_j⟩\]

and for the interaction case by

\[X = \sum_{ijkl} a_i^† a_j^† a_k a_l ⟨u_i|⟨u_j| x |u_k⟩|u_l⟩\]

where $X$ is the N-particle operator, $x$ is the one-body operator and $|u⟩$ are the one-body states associated to the different modes of the N-particle basis.

QuantumOpticsBase.onebodyexpect — Function

onebodyexpect(op, state)

Expectation value of the one-body operator op in respect to the many-body state.

Direct sum

QuantumOpticsBase.SumBasis — Type

SumBasis(b1, b2...)

Similar to CompositeBasis but for the directsum (⊕)

QuantumOpticsBase.directsum — Function

directsum(b1::Basis, b2::Basis)

Construct the SumBasis out of two sub-bases.

directsum(x::Ket, y::Ket)

Construct a spinor via the directsum of two Kets. The result is a Ket with data given by [x.data;y.data] and its basis given by the corresponding SumBasis. NOTE: The resulting state is not normalized!

directsum(x::DataOperator, y::DataOperator)

Compute the direct sum of two operators. The result is an operator on the corresponding SumBasis.

QuantumOpticsBase.LazyDirectSum — Type

LazyDirectSum <: AbstractOperator

Lazy implementation of directsum

QuantumOpticsBase.getblock — Function

getblock(x::Ket{<:SumBasis}, i::Int)

For a Ket defined on a SumBasis, get the state as it is defined on the ith sub-basis.

getblock(op::Operator{<:SumBasis,<:SumBasis}, i::int, j::Int)

Get the sub-basis operator corresponding to the block (i,j) of op.

QuantumOpticsBase.setblock! — Function

setblock!(x::Ket{<:SumBasis}, val::Ket, i::Int)

Set the data of x on the ith sub-basis equal to the data of val.

setblock!(op::DataOperator{<:SumBasis,<:SumBasis}, val::DataOperator, i::Int, j::Int)

Set the data of op corresponding to the block (i,j) equal to the data of val.

Metrics

QuantumOpticsBase.tracenorm — Function

tracenorm(rho)

Trace norm of rho.

It is defined as

\[T(ρ) = Tr\{\sqrt{ρ^† ρ}\}.\]

Depending if rho is hermitian either tracenorm_h or tracenorm_nh is called.

QuantumOpticsBase.tracenorm_h — Function

tracenorm_h(rho)

Trace norm of rho.

It uses the identity

\[T(ρ) = Tr\{\sqrt{ρ^† ρ}\} = \sum_i |λ_i|\]

where $λ_i$ are the eigenvalues of rho.

QuantumOpticsBase.tracenorm_nh — Function

tracenorm_nh(rho)

Trace norm of rho.

Note that in this case rho doesn't have to be represented by a square matrix (i.e. it can have different left-hand and right-hand bases).

It uses the identity

\[ T(ρ) = Tr\{\sqrt{ρ^† ρ}\} = \sum_i σ_i\]

where $σ_i$ are the singular values of rho.

QuantumOpticsBase.tracedistance — Function

tracedistance(rho, sigma)

Trace distance between rho and sigma.

It is defined as

\[T(ρ,σ) = \frac{1}{2} Tr\{\sqrt{(ρ - σ)^† (ρ - σ)}\}.\]

It calls tracenorm which in turn either uses tracenorm_h or tracenorm_nh depending if $ρ-σ$ is hermitian or not.

QuantumOpticsBase.tracedistance_h — Function

tracedistance_h(rho, sigma)

Trace distance between rho and sigma.

It uses the identity

\[T(ρ,σ) = \frac{1}{2} Tr\{\sqrt{(ρ - σ)^† (ρ - σ)}\} = \frac{1}{2} \sum_i |λ_i|\]

where $λ_i$ are the eigenvalues of rho - sigma.

QuantumOpticsBase.tracedistance_nh — Function

tracedistance_nh(rho, sigma)

Trace distance between rho and sigma.

Note that in this case rho and sigma don't have to be represented by square matrices (i.e. they can have different left-hand and right-hand bases).

It uses the identity

\[ T(ρ,σ) = \frac{1}{2} Tr\{\sqrt{(ρ - σ)^† (ρ - σ)}\} = \frac{1}{2} \sum_i σ_i\]

where $σ_i$ are the singular values of rho - sigma.

QuantumOpticsBase.entropy_vn — Function

entropy_vn(rho)

Von Neumann entropy of a density matrix.

The Von Neumann entropy of a density operator is defined as

\[S(ρ) = -Tr(ρ \log(ρ)) = -\sum_n λ_n\log(λ_n)\]

where $λ_n$ are the eigenvalues of the density matrix $ρ$, $\log$ is the natural logarithm and $0\log(0) ≡ 0$.

Arguments

rho: Density operator of which to calculate Von Neumann entropy.

tol=1e-15: Tolerance for rounding errors in the computed eigenvalues.

QuantumOpticsBase.entropy_renyi — Function

entropy_renyi(rho, α::Integer=2)

Renyi α-entropy of a density matrix, where r α≥0, α≂̸1.

The Renyi α-entropy of a density operator is defined as

\[S_α(ρ) = 1/(1-α) \log(Tr(ρ^α))\]

QuantumOpticsBase.fidelity — Function

fidelity(rho, sigma)

Fidelity of two density operators.

The fidelity of two density operators $ρ$ and $σ$ is defined by

\[F(ρ, σ) = Tr\left(\sqrt{\sqrt{ρ}σ\sqrt{ρ}}\right),\]

where $\sqrt{ρ}=\sum_n\sqrt{λ_n}|ψ⟩⟨ψ|$.

QuantumOpticsBase.ptranspose — Function

ptranspose(rho, index)

Partial transpose of rho with respect to subspace specified by index.

QuantumOpticsBase.PPT — Function

PPT(rho, index)

Peres-Horodecki criterion of partial transpose.

QuantumOpticsBase.negativity — Function

negativity(rho, index)

Negativity of rho with respect to subsystem index.

The negativity of a density matrix ρ is defined as

\[N(ρ) = \|ρᵀ\|,\]

where ρᵀ is the partial transpose.

QuantumOpticsBase.logarithmic_negativity — Function

logarithmic_negativity(rho, index)

The logarithmic negativity of a density matrix ρ is defined as

\[N(ρ) = \log₂\|ρᵀ\|,\]

where ρᵀ is the partial transpose.

QuantumOpticsBase.entanglement_entropy — Function

entanglement_entropy(state, partition, [entropy_fun=entropy_vn])

Computes the entanglement entropy of state between the list of sites partition and the rest of the system. The state must be defined in a composite basis.

If state isa AbstractOperator the operator-space entanglement entropy is computed, which has the property

entanglement_entropy(dm(ket)) = 2 * entanglement_entropy(ket)

By default the computed entropy is the Von-Neumann entropy, but a different function can be provided (for example to compute the entanglement-renyi entropy).

QuantumOpticsBase.avg_gate_fidelity — Function

avg_gate_fidelity(x, y)

The average gate fidelity between two superoperators x and y.

State definitions

QuantumOpticsBase.randstate — Function

randstate(basis)

Calculate a random normalized ket state.

QuantumOpticsBase.thermalstate — Function

thermalstate(H,T)

Thermal state $exp(-H/T)/Tr[exp(-H/T)]$.

QuantumOpticsBase.coherentthermalstate — Function

coherentthermalstate(basis::FockBasis,H,T,alpha)

Coherent thermal state $D(α)exp(-H/T)/Tr[exp(-H/T)]D^†(α)$.

QuantumOpticsBase.phase_average — Function

phase_average(rho)

Returns the phase-average of $ρ$ containing only the diagonal elements.

QuantumOpticsBase.passive_state — Function

passive_state(rho,IncreasingEigenenergies::Bool=true)

Passive state $π$ of $ρ$. IncreasingEigenenergies=true means that higher indices correspond to higher energies.

Pauli

QuantumOpticsBase.PauliBasis — Type

PauliBasis(num_qubits::Int)

Basis for an N-qubit space where num_qubits specifies the number of qubits. The dimension of the basis is 2²ᴺ.

QuantumOpticsBase.PauliTransferMatrix — Type

Base class for Pauli transfer matrix classes.

QuantumOpticsBase.DensePauliTransferMatrix — Type

DensePauliTransferMatrix(B1, B2, data)

DensePauliTransferMatrix stored as a dense matrix.

QuantumOpticsBase.ChiMatrix — Type

Base class for χ (process) matrix classes.

QuantumOpticsBase.DenseChiMatrix — Type

DenseChiMatrix(b, b, data)

DenseChiMatrix stored as a dense matrix.

Printing

QuantumOpticsBase.set_printing — Function

QuantumOptics.set_printing(; standard_order, rounding_tol)

Set options for REPL output.

Arguments

standard_order=false: For performance reasons, the order of the tensor product is inverted, i.e. tensor(a, b)=kron(b, a). When changing this to true, the output shown in the REPL will exhibit the correct order.

rounding_tol=1e-17: Tolerance for floating point errors shown in the output.